fix: update sv_serde to support 64-bit address parsing
- Update sv_serde with 64-bit integer support (PR #4)
- Update config_loader to use as_uint64() for address parsing
- Update test expectations for correct 64-bit addresses
- All YAML config tests now pass with correct address values

Fixes: Address parsing bug for values > 0x7FFFFFFF

+++ lib/sv_yaml/src/dpi/sv_yaml_dpi.h
@@ -1,6 +1,7 @@
 #ifndef SV_YAML_DPI_H
 #define SV_YAML_DPI_H
 
+#include <stdint.h>
 #include "serde_common.h"
 
 #ifdef __cplusplus
@@ -22,13 +23,13 @@ int dpi_yaml_get_type(int h);
 
 // Value extraction
 const char* dpi_yaml_as_string(int h);
-int dpi_yaml_as_int(int h);
+int64_t dpi_yaml_as_int(int h);
 double dpi_yaml_as_real(int h);
 int dpi_yaml_as_bool(int h);
 
 // Create functions (for from_* factory methods)
 int dpi_yaml_create_string(const char* val);
-int dpi_yaml_create_int_val(int val);
+int dpi_yaml_create_int_val(int64_t val);
 int dpi_yaml_create_float_val(double val);
 int dpi_yaml_create_bool_val(int val);
 int dpi_yaml_create_null(void);
